WheelPicker - move from incubator (#2486)
