+++ src/incubator/index.ts
@@ -5,7 +5,5 @@ export {default as ExpandableOverlay} from './expandableOverlay';
 export {default as TextField, TextFieldProps, FieldContextType, TextFieldMethods} from './TextField';
 export {default as Toast, ToastProps, ToastPresets} from './toast';
 export {default as TouchableOpacity, TouchableOpacityProps} from './TouchableOpacity';
-export {default as WheelPicker, WheelPickerProps, WheelPickerItemProps} from './WheelPicker';
-export {WheelPickerAlign} from './WheelPicker/types';
 export {default as PanView, PanViewProps, PanViewDirections, PanViewDismissThreshold} from './panView';
 export {default as Dialog, DialogProps, DialogHeaderProps, DialogStatics, DialogImperativeMethods} from './Dialog';
